+++ lce/src/main/java/com/laimiux/lce/CE.kt
@@ -8,7 +8,6 @@ interface CE<out C, out E> {
     companion object {
         fun <T> content(content: T): CE<T, Nothing> = Type.Content(content)
 
-        fun error(error: Throwable): CE<Nothing, Throwable> = Type.Error(error)
         fun <T> error(error: T): CE<Nothing, T> = Type.Error(error)
 
         /**