+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 ## [x.x.x] - TBD
-- 
+- Added `UC.merge()` extension method
 
 ## [0.2.0] - January 11, 2021 
 - Added LC type.