+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [x.x.x] - TBD
 - Added `UC.merge()` extension method
 - Added `LC.merge()` extension method
+- Added `UCT.asUC { throwable -> UC.loading }` extension method
 
 ## [0.2.0] - January 11, 2021 
 - Added LC type.