+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added `doOnContent` extension to `Observable<CE>` type.
 - Added `flatten` extension method to `LC` and `UC` types.
 - Added `asUCT()` extension to `UCE<C, Throwable>` type.
+- Remove `LCE.error(throwable)`, `UCE.error(throwable)` and `CE.error(throwable)`
 
 ## [0.2.0] - January 11, 2021 
 - Added LC type.